[Commit 10] First Playable Commit
This encompasses all of the progress submitted for our first playable. I
shifted the depth on collision and teleport objects to -1000 so that
they're always shown on top in the room editor and I converted the win
trigger into a general purpose message trigger and  added a temporary
knockback on the player when they collide with it so they don't get
stuck.