[Commit 12] Art Commit
Added several art assets and shift some functionality in an attempt to
reconcile the art with gameplay. Still have plenty that needs to be
done, at this point there is far more art available than is in engine
and I'm trying to get it all in so that we can work on getting the art
and gameplay to play nice with each other.